requirements.txt
Pillow
numpy
scipy
opencv-python